Update patch set 5
Patch Set 5:

If you were to use the directory structure to represent the scenario ('fuel-os-odl_l2-bgpvpn' instead of 'fuel') the eventual document would be posted as a link link opnfv://installation/os-odl_l2-bgpvpn/index.rst.
I would suggest using the scenario name as the directory in order that this is not eventually confused with other fuel related documents.

Patch-set: 5
Label: Verified=0