DO NOT MERGE Fix network assoc tests in Boron
Network assoc in Boron with netvirt does not work for subnets
that have a router due to an inherent limitation in the implementation.
See the mailing list and related bug:
https://bugs.opendaylight.org/show_bug.cgi?id=6962
https://lists.opendaylight.org/pipermail/netvirt-dev/2016-October/001815.html
There seems to also be an issue with route distinguishers. Creating a
network assoc in a bgpvpn that does not have a route distinguisher does
not work.
This commit fixes the net assoc case and works around these issues by:
- Removing the routers from the subnets in testcase_1
- Adding a route distinguisher to the bgpvpn

Change-Id: I6b57eab89839d9e9122cd24b0f05737467439dd9
Signed-off-by: Romanos Skiadas <rski@intracom-telecom.com>