Fix network assoc tests in Boron
Network assoc in Boron with netvirt does not work for subnets that have
a router due to an inherent limitation in the implementation. See the
mailing list and related bug:
https://bugs.opendaylight.org/show_bug.cgi?id=6962
https://lists.opendaylight.org/pipermail/netvirt-dev/2016-October/001815.html
Also, if a subnet does not have a route distinguisher, it is not added
to the FIB, so association does not make traffic move between subnets.
This is intentional and an error is logged when a subnet is associated
to a bvpvpn without a route distinguisher.

This commit fixes the net assoc case and works around these issues by:
- Removing the routers from the subnets in testcase_1
- Adding a route distinguisher to the bgpvpn

JIRA: SDNVPN-74
Change-Id: I6b57eab89839d9e9122cd24b0f05737467439dd9
Signed-off-by: Romanos Skiadas <rski@intracom-telecom.com>
(cherry picked from commit b9eb7024b014cba0d299b1cf3b01e179c7d0482e)