Update patch set 5
Patch Set 5:

Well, it's only 5 instances, and each one of them have different parameters. I would prefer to not use heat for now, but it's an option. Maybe there is a smart way to put that creation in a for loop and put the parameters of each VM in the config file somehow.

Patch-set: 5
Reviewer: Gerrit User 459 <459@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0