Update patch set 5
Patch Set 5: Code-Review-1

Hey,

I am sorry to step in but that was not really the intend of my comments on the last patchsets. as you see we do not gain here much since the config for all testcases is still very long.

I thought mor about function then config.yaml. which generate then the right config. so that testcase 1 and two just through a flag can tell how many vms they want to have etc.

Now that the config is just sitting in another format in another file the users will just be more confused or?

Patch-set: 5