Update patch set 4
Patch Set 4:

One more thing. This script is run from the jumphost and not from the undercloud vm anymore. That is great actually. I wanted to use ironic, nova and neutron client but I think your idea of getting the ips is far more easy and with that more robust. So the script deployment_cloner.py is as well not needed anymore.

Patch-set: 4