Exchange routes from ODL to Quagga- testcase3
This is an extension of testcase3.
After the peering between zrpcd (ODL side) and Quagga,
a network defined on ODL/zpcd side is advertised towards
vanilla Quagga.

Initially a vrf with route-distinguisher and route targets
are defined.

For the time being the Quagga accepts a new prefix but cannot
understand the value of it. This is going to be accomplished,
defining a vrf set on this side too.

JIRA: SDNVPN-112

Change-Id: I9ba677e74f24258f7cc59db70b013fbdbbec917a
Signed-off-by: tomsou <soth@intracom-telecom.com>