Update patch set 1
Patch Set 1:

It's beteer to have patches which target single features.
I would change the tittle to "Add test case 9", since it's the main purpose of this, isn't it? And if the fail mode is a new feature, it could be in a separated patch with dependencies if you want.

Patch-set: 1
Label: Code-Review=0