Add/Remove testcase
This testcase investigates if communication
between a group of VMs is interupted upon
deleteon and creation of VMs inside this group.
Test case flow:
3 vms:
[1] on compute 1,
[2] on compute 1,
[3] on compute 2,
all vms ping each other.

vm [2] is deleted. Traffic is still flying
between [1] and [3]. A new [4] vm is added
to compute 1. Traffic is not interrupted
and [4] can be reached as well.

JIRA: SDNVPN-105

Change-Id: I93a65de2a5df83551e3115fd20b60241cd460e97
Signed-off-by: panageo2 <panageo@intracom-telecom.com>