Test group_add_mod remove ODL error in ovs group
Create a new testcase (Testcase 11)

Testing procedure:
Create a Network with 2 VMs so that a group is added
to ovs. Disconnect ovs. Remove the Network and the VMs.
Connect ovs again. Check that the group is deleted from
it and the logs don't have the error group already
exists.

JIRA: SDNVPN-151

Change-Id: I4c6f5e16f837bbcc589cbbbf05896a2d17b9cd57
Signed-off-by: panageo2 <panageo@intracom-telecom.com>