Update patch set 2
Patch Set 2:

> (1 comment)
 > 
 > good please continue :-D

 > (1 comment)
 > 
 > good please continue :-D

Hello Nikolas!

This file is actually copied at 211 220 lines of testcase_2.py. If we do not have this if then 

 > (1 comment)
 > 
 > good please continue :-D

Hello Nikolas!

This file is actually copied at lines 211 220 of testcase_2.py. If we do not place this if here and the file has not been copied (the script runs at the initialization of instance before the id_rsa file is copied) then at line 211 of utils.py it ssh will prompt for password and the script will hang waiting for password input. we should not create the id_rsa file in an else. we should wait for it to be copied.

Patch-set: 2