Update patch set 3
Patch Set 3:

I don't think you need to call this function since the framework will do it for you. You only need to populate the varaible
    self.details
with some formatting, for instance:
    'details': [
    {u"'sdnvpn-1-1' can ping 'sdnvpn-1-2'": 'PASS'},
    {u"'sdnvpn-1-1' can ping 'sdnvpn-1-3'": 'PASS'},
    {u"'sdnvpn-1-1' cannot ping 'sdnvpn-1-4'": 'PASS'},
    {u"'sdnvpn-1-4' can ping 'sdnvpn-1-5'": 'PASS'},
    {u"'sdnvpn-1-1' can ping 'sdnvpn-1-4'": 'PASS'},
    {u"'sdnvpn-1-1' can ping 'sdnvpn-1-5'": 'PASS'}]
or similar. The format for the "details" argument is 100% free, it can be anything as long as it follows some logic.

Patch-set: 3