Update patch set 2
Patch Set 2: Code-Review+1

This is a good approach, however it won't work until we adapt how Functest calls SDNVPN.

Currently, Functest executes the script run_sdnvpn_tests.py as a bash command, see [1]. Therefore, self.details is not part of the class in Functest framework and will be useless. 

What we need to do after Euphrates is to run bgpvpn test as a python module (see how it is done for Barometer). This way we can make use of all internal things that the class Testcase offers in Functest.

[1] https://git.opnfv.org/functest/tree/docker/features/testcases.yaml#n54

Patch-set: 2
Label: Code-Review=0