Update patch set 1
Patch Set 1: Code-Review-1

I would like to suggest a different approach. The objective of this test is to check if the VRF for a VPN gets set up correctly depending on the RT config. We can achieve that by deleting and re-creating the VPNs instead of updating the RTs. We should create a second testcase specifically isolating the RT update problem and disable that one for the time being.

Patch-set: 1
Reviewer: Gerrit User 321 <321@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=-1