Intra-Data Center ECMP testcase
This testcase add a scenario in which one extra route can be reachable
through two VMs which are in a same network attached to a router which
is later associated to a BGPVPN instance.
This extra route ip address is different from subnet of VMs.

After the extra route is configured with VM ips as nexthops, then when
third VM is trying to reach extra route, traffic would split between two
VMs. The traffic split can happen only in heavy loaded scenario, so this
testcase would test only reachability.

JIRA: SDNVPN-23

Change-Id: I4bb4e94f5c98c0f3c22bff7a6998af2a7a919f16
Signed-off-by: Periyasamy Palanisamy <periyasamy.palanisamy@ericsson.com>