Update patch set 5
Patch Set 5: Code-Review+1

This is good. However, be careful with one thing: OPNFV is going to release Fraser based on the openstack Pike version. That means, openstack-ansible will also be in the Pike version and thus, the bgp-vpn integration in xci is not there. We are fixing this is xci by having two different a-r-r files, one pointing to master and the other one pointing to pike SHAs except for os_neutron that points to a private repo we have where the sfc related logic was cherry picked into the pike branch. Perhaps you could do something similar and then, your tasks/main.yml can check the OSA_version variable to decide if it copies the a-r-r file

Patch-set: 5
Label: Code-Review=+1