remove block comment for testcase_8
somehow bgpvpn testsuite is not running in the bgpvpn-ha ci loop
because of testcase_8. when bgpvpn tries to execute this testcase,
somehow execution is stopped and no logs being generated from the
setup. The only different is this testcase is having block comment
with delimeter and this is being imported as a module inside
run_sdnvpn_tests. Hence replacing block comment with single line
comments. This issue is very sporadic even in the local setup
and running fine after changing it with single line comments.

Change-Id: I5d71ec2a8c532b5dd77b2136ca7e95a7838683dc
Signed-off-by: Periyasamy Palanisamy <periyasamy.palanisamy@ericsson.com>