Create and pass references of flavor and image to tempest
When invoked from SDNVPN, Tempest scenario functest test cases
try to use flavor and image references from the tempest.conf file
which are currently empty.
This patch uses the default flavor that is created by the test
suite and the creates an image using the availabe cirros image,
which is deleted at the end.

JIRA: SDNVPN-209

Change-Id: I6a6fd636fec2c0fa7d35981ca51d59f06a4d9d69
Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>