[Bug Fix] Determine the BGP entity owner in ODL Cluster
In HA deployments is essential to determine which is the ODL
BGP entity owner and transform that ODL to BGP speaker.
This is necessary because of this bug [0] in netvirt. So
we can consider this as a workaround for the testcase_3.

Note: This patch is dependent on this task resolution [1]
      for master branch.

Note: This patch works fine in stable/fraser branch.

[0] https://jira.opendaylight.org/browse/NETVIRT-1308
[1] https://jira.opnfv.org/browse/SDNVPN-217

Jira: SDNVPN-214

Change-Id: I383987e650da45400a3a6437dbdaaea904db9265
Signed-off-by: Dimitrios Markou <mardim@intracom-telecom.com>