Update patch set 1
Patch Set 1:

This might need some more debugging. 
I just tried to with CLI to associate floating ip address with neutron port if its router is already set with external network gateway. it worked.

We are trying same thing in the testcase. But not sure why this doesn't work.

Patch-set: 1
Label: Verified=0