Adds support for docker based opendaylight
Previously we used ODL as a systemd service. However in newer OOO it is
ran as a docker container. For reinstalling ODL in this case we determined
that the best way is to not mess with docker and simply shut it down,
and use the systemd for the CSIT verification.

This patch stops the docker container, then copies a systemd service
file (since RPM may never have been installed), and then allows ODL
reinstall to proceed as usual.

Change-Id: Ib0996ed5494f1ef37bcc232f087f563c99db62fc
Signed-off-by: Tim Rozet <trozet@redhat.com>