Update patch set 2
Patch Set 2: Code-Review+2

I actually don't think it is a good idea to install it in a different way then it was installed before on the system. There is a good reason why it runs in a docker container I guess. And what if the java version etc changes? Should be less error prone when using docker. However I guess you already tried and run into some difficulties, so this is the working solution now.

Patch-set: 2
Label: Code-Review=+2