Update patch set 2
Patch Set 2:

> I actually don't think it is a good idea to install it in a
 > different way then it was installed before on the system. There is
 > a good reason why it runs in a docker container I guess. And what
 > if the java version etc changes? Should be less error prone when
 > using docker. However I guess you already tried and run into some
 > difficulties, so this is the working solution now.

It's a valid comment. Ideally we do want to use do what you suggested, but as you also mentioned we looked at that and it is going to be more complex because we do not have the undercloud anymore to help us setup the container. So as a first step we will try it this way and then hopefully have another patch later to do it with docker. But for now, our main concern is getting a working ODL up.

Patch-set: 2
Label: Verified=0