Update patch set 1
Patch Set 1:

Based on my experience cleaning up systems manually using the CLI, it might not be enough to just try to delete the routers right away. For instance, I often need to unset the gateway of the router and remove the interfaces to the subnets. Pkease keep an eye on this.

Patch-set: 1
Reviewer: Gerrit User 585 <585@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0