Update patch set 2
Patch Set 2:

> I don't see why you should override method here.
 > e.g. services are well discovered

Hi Cedric,

Thank you once again for the quick reply. I think that you are referring to the override of the *configure* method. Indeed, the image and the flavor are unnecessary and I will remove them.

But concerning the services part I think there is a problem. I make the assumption that by services you mean bgpvpn neutron_extension. Indeed in the master branch of functest I can see a method in TempestCommon which checks the neutron_extensions found in the testcases.yaml, so I suppose the update of the yaml in this patch would be enough for master. Although, I cannot see that method in the stable/fraser branch of functest.

With this patch I am trying to run tempest test-cases for both stable/fraser and master branches. When I tried to run it in stable/fraser without overriding *configure* to add bgpvpn=True to services_available, all test-cases were skipped. After override, all test-cases ran.

Is there any other alternative in order to run the test-cases for both stable/fraser and master with the same code?

Thank you in advance.

Kind regards,
Stamatis

Patch-set: 2