Fixes race condition with checking OF controller
There have been cases where the OF controller check to ensure ODL
configured the right URL, as well as checking the OVS is connected via
OpenFlow to ODL has failed due to a race condition where ODL has just
set the controller URL in OVS, but has not fully connected yet. This
patch changes the code to poll multiple times.

Change-Id: I24704d3f4e6879dd8a8a7079f45c6cf1473f5bbe
Signed-off-by: Tim Rozet <trozet@redhat.com>