Using heat orchestrator for sdnvpn - test case 4
Heat orchestrator and the use of Heat Orchestrator Templates
is introduced in sdnvpn test cases. The deployment of the
nodes and networks under test is performed as a stack with the
use of the openstack api and the use of the other apis is kept
to as little as possible. The scenarios that are executed are
the same as in the orginal test cases.

This is the implementation of sdnvpn test case 4:
VPN provides connectivity between subnets using router
association

JIRA: SDNVPN-219

Change-Id: If3783dd25577d22c18be5f088ab6657c38531d73
Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>