Using heat orchestrator for sdnvpn - test case 2
Heat orchestrator and the use of Heat Orchestrator Template
is introduced in sdnvpn test cases. The deployment of the
nodes and networks under test is performed as a stack with the
use of the openstack api and the use of the other apis is kept
to as little as possible. The scenarios that are executed are
the same as in the orginal test cases.

This is the implementation of sdnvpn test case 2:
Tenant separation

JIRA: SDNVPN-219

Change-Id: I20f5ed8f7d250873f0223c1cd00c6c05d00fbfa8
Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>