Using heat orchestrator for sdnvpn - test case 8
Heat orchestrator and the use of Heat Orchestrator Templates
is introduced in sdnvpn test cases. The deployment of the
nodes and networks under test is performed as a stack with the
use of the heat api and the use of the other apis is kept to
as little as possible. The scenarios that are executed are the
same as in the orginal test cases.

This is the implementation of sdnvpn test case 8:
Test floating IP and router assoc coexistence

JIRA: SDNVPN-219

Change-Id: I8a61d40e9decf4297b8714c93ce77ba05f033b66
Signed-off-by: nikoskarandreas <nick@intracom-telecom.com>